a64_linux: update to orange-pi-6.1-20230214-2103
This commit updates the Linux kernel used for A64 driver support to
the latest orange-pi-6.1 branch, which corresponds to version 6.1.12.

Fixes #21.

+++ src/drivers/framebuffer/de/fb.c
@@ -20,7 +20,7 @@ struct fb_info * framebuffer_alloc(size_t size,struct device * dev)
 {
 #define BYTES_PER_LONG (BITS_PER_LONG/8)
 #define PADDING (BYTES_PER_LONG - (sizeof(struct fb_info) % BYTES_PER_LONG))
-	int fb_info_size = sizeof(struct fb_info);
+	size_t fb_info_size = sizeof(struct fb_info);
 	struct fb_info *info;
 	char *p;
